Working on better character selection

+++ Source/KrzyweKarty/Cards/KKCharacter.h
@@ -154,7 +154,7 @@ public:
 	// Tiles
 
 	UFUNCTION(BlueprintCallable)
-	virtual int32 GetTilePositionID() const override;
+	virtual uint8 GetTilePositionID() const override;
 
 	virtual void OnSelectableHighlighted() override;
